initialize slash-create and SSE event listeners similar to discordjs events, resturcture and refactor, add vote ephemeral user response
